+++ src/CSGOMatchData.h
@@ -5,6 +5,8 @@
 #include <steam/steamtypes.h>
 #include <vector>
 #include <string>
+#include <fmt/format.h>
+#include <fmt/color.h>
 
 struct CSGOMatchData
 {
@@ -23,11 +25,11 @@ struct CSGOMatchData
         uint64 reservation_id;
 
         std::string sharecode;
-        std::string demolink;   /* roundstats.map */
+        std::string replaylink;   /* roundstats.map */
 
         std::string map;        /* watchablematchinfo.game_map */
         std::string mapgroup;   /* watchablematchinfo.game_mapgroup */
-        uint32 gametype;   /* roundstatsall.reservation.game_type | watchablematchinfo.game_type */
+        uint32 gametype;        /* roundstatsall.reservation.game_type | watchablematchinfo.game_type */
 
         uint32 spectators;
 
@@ -38,6 +40,30 @@ struct CSGOMatchData
 
         int score_ally;
         int score_enemy;
+
+        std::string getScore() const
+        {
+          return fmt::format("{:02} : {:02}", score_ally, score_enemy);
+        }
+
+        // TODO how to get mapname?
+        std::string getMapname() const
+        {
+          return (map.empty() ? "? " : map);
+        }
+
+        std::string getMatchResult() const
+        {
+          if(result_str == "LOSS") {
+            return fmt::format(fmt::fg(fmt::color::red), "LOSS");
+          }
+          else if(result_str == "WIN") {
+            return fmt::format(fmt::fg(fmt::color::green), "WIN");
+          }
+          else { // result_str == "TIE"
+            return fmt::format(fmt::fg(fmt::color::yellow), "TIE");
+          }
+        }
 };
 
 #endif